Allow output_dir to be specified on the command line
The output dir can now be passed in on the command line. The script
simply takes the first parameter on the command line and sets that to
the output_dir. If no option is passed it will use the temp directory.

This can be used with a web server to generate the rdocs into a
publically viewable directory to be served. Sorta ironic since the point
of bdoc was to remove the server from the equation.